refactor(cli-engine): restructure execution around an EngineImpl class
run.ts and invocation.ts are gone. execution/engine.ts holds the
Engine interface and an EngineImpl class whose methods are the execute
pipeline; buildEngine constructs one. createCli lives in src/cli.ts
with the Cli interface, and the test harness implementation in
src/testing.ts. Every settle* function, including settleVersion and
settleUnhandled, now lives in settlement.ts.

Modules are named for their concept, not their mechanics:
execution/events.ts (ctx.report filtering + json framing) is
reporting.ts, render.ts is rendering.ts, and firstLine joins the
rendering helpers.
