+++ packages/cli-engine/src/execution/settlement.ts
@@ -1,10 +1,19 @@
-import type { AnyCommand } from "../commands";
-import type { CompletedEnvelope, ErroredEnvelope } from "../envelopes";
+import type {
+  AnyCommand,
+  CompletedEnvelope,
+  ErroredEnvelope,
+} from "../commands";
 import { PRESENTED, type PresentedResult } from "../presentation";
 import { CliStructuredError, type Diagnostic } from "../protocol";
-import { emitFrame } from "./events";
-import { firstLine, type Invocation } from "./invocation";
-import { renderCompletedHuman, withDocsUrl, writeDiagnostic } from "./render";
+import type { EngineSpec, Invocation } from "./engine";
+import {
+  firstLine,
+  renderCompletedHuman,
+  withDocsUrl,
+  writeDiagnostic,
+} from "./rendering";
+import { emitFrame } from "./reporting";
+import { usageErrorCode } from "./stricli-adapter";
 
 function undocumentedExitCode(
   def: AnyCommand,
@@ -14,9 +23,7 @@ function undocumentedExitCode(
     return undefined;
   }
   const documented =
-    def.kind === "result-command" && def.exitCodes !== undefined
-      ? Object.keys(def.exitCodes).map(Number)
-      : [];
+    def.kind === "result-command" ? Object.keys(def.exitCodes).map(Number) : [];
   if (documented.includes(exitCode)) {
     return undefined;
   }
@@ -63,7 +70,7 @@ export function settleCompleted(
       diagnostics: presented.diagnostics.map((diagnostic) =>
         withDocsUrl(state, diagnostic),
       ),
-      nextActions: presented.presentation.next ?? [],
+      nextActions: presented.presentation.next,
     };
     emitFrame(invocation, {
       kind: "result",
@@ -93,11 +100,11 @@ export function settleErrored(
     commandId: state.commandId,
     error: diagnosticOf(error),
     diagnostics,
-    nextActions: error.nextActions ?? [],
+    nextActions: error.nextActions,
   });
 }
 
-/** Signal exit codes (R6): 130 SIGINT, 143 SIGTERM. The engine's own
+/** Signal exit codes: 130 SIGINT, 143 SIGTERM. The engine's own
  *  controller only ever aborts with a signal name. */
 function signalExitCode(reason: unknown): number {
   return reason === "SIGTERM" ? 143 : 130;
@@ -133,6 +140,7 @@ function settleAborted(invocation: Invocation): void {
       code: "CLI.ABORTED",
       severity: "error",
       summary: "The command was aborted before it completed.",
+      nextActions: [],
     },
     diagnostics: [],
     nextActions: [],
@@ -176,6 +184,7 @@ export function settleBug(invocation: Invocation, cause: unknown): void {
       summary: firstLine(
         cause instanceof Error ? cause.message : String(cause),
       ),
+      nextActions: [],
     },
     diagnostics: [],
     nextActions: [],
@@ -208,4 +217,79 @@ export function emitErrored(
   for (const diagnostic of envelope.diagnostics) {
     writeDiagnostic(stderr, diagnostic);
   }
+}
+
+/** `--version` prints createCli's version and exits 0. In json mode the
+ *  version travels as the run's single result frame. */
+export function settleVersion(
+  spec: EngineSpec,
+  invocation: Invocation,
+): number {
+  const { runtime, state } = invocation;
+  if (state.format === "human") {
+    runtime.stdout.write(`${spec.version}\n`);
+    return 0;
+  }
+  emitFrame(invocation, {
+    kind: "result",
+    envelope: {
+      ok: true,
+      commandId: "version",
+      result: { version: spec.version },
+      exitCode: 0,
+      diagnostics: [],
+      nextActions: [],
+    },
+    commandId: "version",
+    timestamp: invocation.now().toISOString(),
+  });
+  return 0;
+}
+
+/** Maps stricli's own settlement (parse/route failures, framework bugs)
+ *  onto the engine protocol when the pipeline never settled. A failure
+ *  addressed at a server command renders to stderr — a foreign client on
+ *  the other end of stdout must never receive an engine envelope. */
+export function settleUnhandled(
+  spec: EngineSpec,
+  invocation: Invocation,
+  stricliExitCode: number | string | null | undefined,
+): number {
+  const state = invocation.state;
+  const raw = typeof stricliExitCode === "number" ? stricliExitCode : 0;
+  if (raw === 0) {
+    return 0;
+  }
+  const segments =
+    state.prefix[0] === spec.name ? state.prefix.slice(1) : state.prefix;
+  if (spec.commands[segments.join(" ")]?.kind === "server-command") {
+    state.format = "human";
+  }
+  const usage = usageErrorCode(raw) !== undefined;
+  const captured = usage
+    ? state.stricliStderr.trim()
+    : (state.usageErrorText ??
+      state.internalErrorText ??
+      state.stricliStderr.trim());
+  const full =
+    captured.length > 0 ? captured : "The command failed unexpectedly";
+  const summary = firstLine(full);
+  const remainder = full.slice(full.indexOf("\n") + 1).trim();
+  const envelope: ErroredEnvelope = {
+    ok: false,
+    commandId: segments.join("."),
+    error: {
+      code: usageErrorCode(raw) ?? "CLI.INTERNAL_ERROR",
+      severity: "error",
+      summary,
+      ...(usage && full.includes("\n") && remainder.length > 0
+        ? { why: remainder }
+        : {}),
+      nextActions: [],
+    },
+    diagnostics: [],
+    nextActions: [],
+  };
+  emitErrored(invocation, envelope);
+  return usage ? 2 : 1;
 }
